Implemented auto resolve reports on delete note

+++ src/server/queries/delete.ts
@@ -7,7 +7,15 @@ import { notes, reports } from '~/server/db/schema';
 import { utapi } from '~/server/uploadthing';
 
 export async function deleteNote(utId: SelectNote['utId']) {
+  const note = await db.select().from(notes).where(eq(notes.utId, utId)).get();
+
+  if (!note) {
+    throw new Error(`Note with utId ${utId} not found`);
+  }
+
+  await db.update(reports).set({ noteId: null, status: "resolved" }).where(eq(reports.noteId, note.id));
   await utapi.deleteFiles(utId);
+
   return db.delete(notes).where(eq(notes.utId, utId));
 }
